# docs/swagger.json
{"swagger":"2.0","info":{"description":"API for managing systemd services and containers","title":"Sirberus API","contact":{},"version":"1.0"},"basePath":"/api","paths":{"/container":{"get":{"description":"Get a list of all containers","produces":["application/json"],"tags":["containers"],"summary":"List containers","responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ContainerList"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}},"/container/{id}":{"get":{"description":"Get detailed information about a specific container","produces":["application/json"],"tags":["containers"],"summary":"Get container details","parameters":[{"type":"string","description":"Container ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Container"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ErrorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}},"/container/{id}/exec":{"post":{"description":"Execute a command in a container and stream the output","consumes":["application/json"],"produces":["text/event-stream"],"tags":["containers","sse"],"summary":"Execute command in container","parameters":[{"type":"string","description":"Container ID","name":"id","in":"path","required":true},{"description":"Command to execute","name":"command","in":"body","required":true,"schema":{"$ref":"#/definitions/ContainerExecRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SSEvent"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/SSEvent"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/SSEvent"}}}}},"/container/{id}/logs":{"get":{"description":"Stream logs from a container (always includes real-time updates)","produces":["text/event-stream"],"tags":["containers","sse"],"summary":"Stream container logs","parameters":[{"type":"string","description":"Container ID","name":"id","in":"path","required":true},{"type":"integer","default":100,"description":"Number of historical log lines to return before streaming new ones","name":"lines","in":"query"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SSEvent"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/SSEvent"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/SSEvent"}}}}},"/container/{id}/restart":{"post":{"description":"Restart a container","produces":["application/json"],"tags":["containers"],"summary":"Restart container","parameters":[{"type":"string","description":"Container ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Message"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ErrorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}},"/container/{id}/start":{"post":{"description":"Start a container","produces":["application/json"],"tags":["containers"],"summary":"Start container","parameters":[{"type":"string","description":"Container ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Message"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ErrorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}},"/container/{id}/stop":{"post":{"description":"Stop a container","produces":["application/json"],"tags":["containers"],"summary":"Stop container","parameters":[{"type":"string","description":"Container ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Message"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ErrorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}},"/systemd":{"get":{"description":"Get a list of all systemd services","produces":["application/json"],"tags":["systemd"],"summary":"List systemd services","responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SystemdServiceList"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}},"/systemd/{name}":{"get":{"description":"Get detailed information about a specific systemd service","produces":["application/json"],"tags":["systemd"],"summary":"Get systemd service details","parameters":[{"type":"string","description":"Service name","name":"name","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SystemdServiceDetails"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ErrorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}},"/systemd/{name}/logs":{"get":{"description":"Stream logs from a systemd service (always includes real-time updates)","produces":["text/event-stream"],"tags":["systemd","sse"],"summary":"Stream service logs","parameters":[{"type":"string","description":"Service name","name":"name","in":"path","required":true},{"type":"integer","default":100,"description":"Number of historical log lines to return before streaming new ones","name":"lines","in":"query"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SSEvent"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/SSEvent"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/SSEvent"}}}}},"/systemd/{name}/restart":{"post":{"description":"Restart a systemd service","produces":["application/json"],"tags":["systemd"],"summary":"Restart service","parameters":[{"type":"string","description":"Service name","name":"name","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Message"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ErrorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}},"/systemd/{name}/start":{"post":{"description":"Start a systemd service","produces":["application/json"],"tags":["systemd"],"summary":"Start service","parameters":[{"type":"string","description":"Service name","name":"name","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Message"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ErrorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}},"/systemd/{name}/stop":{"post":{"description":"Stop a systemd service","produces":["application/json"],"tags":["systemd"],"summary":"Stop service","parameters":[{"type":"string","description":"Service name","name":"name","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Message"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ErrorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}}},"definitions":{"Container":{"type":"object","properties":{"command":{"description":"Command running in the container","type":"string"},"cpuUsage":{"description":"CPU usage in nanoseconds","type":"number"},"created":{"description":"Creation time","type":"string"},"environment":{"description":"Container environment variables","type":"array","items":{"type":"string"}},"id":{"description":"Short container ID","type":"string"},"image":{"description":"Container image","type":"string"},"isRunning":{"description":"Whether the container is currently running","type":"boolean"},"labels":{"description":"Container labels","type":"object","additionalProperties":{"type":"string"}},"memoryUsage":{"description":"Memory usage in bytes","type":"integer"},"mounts":{"description":"Container mount points","type":"array","items":{"$ref":"#/definitions/Mount"}},"name":{"description":"Container name","type":"string"},"networks":{"description":"Container network configurations","type":"object","additionalProperties":{"$ref":"#/definitions/NetworkConfig"}},"ports":{"description":"Exposed ports","type":"string"},"size":{"description":"Container size","type":"string"},"status":{"description":"Container status","type":"string"}}},"ContainerExecRequest":{"type":"object","properties":{"command":{"description":"Command to execute","type":"string"}}},"ContainerList":{"type":"object","properties":{"containers":{"description":"List of containers","type":"array","items":{"$ref":"#/definitions/Container"}},"count":{"description":"Total count of containers","type":"integer"}}},"ErrorResponse":{"type":"object","properties":{"error":{"type":"string"}}},"Message":{"type":"object","properties":{"message":{"type":"string"}}},"Mount":{"type":"object","properties":{"destination":{"description":"Destination path in container","type":"string"},"mode":{"description":"Mount mode (ro, rw)","type":"string"},"source":{"description":"Source path on host","type":"string"}}},"NetworkConfig":{"type":"object","properties":{"gateway":{"description":"Network gateway","type":"string"},"ipAddress":{"description":"Network IP address","type":"string"},"macAddress":{"description":"Network MAC address","type":"string"}}},"SSEvent":{"type":"object","properties":{"content":{}}},"SystemdService":{"type":"object","properties":{"activeState":{"description":"Active state (e.g., \"active\", \"inactive\")","type":"string"},"cpuUsage":{"description":"CPU usage as percentage of a single core (can exceed 100% if using multiple cores)","type":"number"},"description":{"description":"Service description","type":"string"},"loadState":{"description":"Load state (e.g., \"loaded\", \"not-found\")","type":"string"},"memoryUsage":{"description":"Memory usage in bytes","type":"integer"},"name":{"description":"Service name","type":"string"},"subState":{"description":"Sub state (e.g., \"running\", \"dead\")","type":"string"},"uptime":{"description":"Uptime in seconds (time since service was started)","type":"integer"}}},"SystemdServiceDetails":{"type":"object","properties":{"cGroup":{"description":"Control group path","type":"string"},"cpuTimeNSec":{"description":"CPU time in nanoseconds","type":"integer"},"docs":{"description":"Documentation URLs","type":"array","items":{"type":"string"}},"dropIn":{"description":"Drop-in configuration paths","type":"array","items":{"type":"string"}},"fragmentPath":{"description":"Path to the unit file","type":"string"},"invocation":{"description":"Invocation ID","type":"string"},"ioReadBytes":{"description":"Bytes read from disk","type":"integer"},"ioWriteBytes":{"description":"Bytes written to disk","type":"integer"},"ipEgressBytes":{"description":"Bytes sent over IP","type":"integer"},"ipIngressBytes":{"description":"Bytes received over IP","type":"integer"},"mainPID":{"description":"Main process ID","type":"integer"},"mainProcess":{"description":"Main process command","type":"string"},"memoryPeak":{"description":"Peak memory usage in bytes","type":"integer"},"processes":{"description":"List of processes","type":"array","items":{"type":"string"}},"service":{"$ref":"#/definitions/SystemdService"},"since":{"description":"Timestamp when the service was activated (RFC3339 format)","type":"string"},"tasks":{"description":"Current number of tasks","type":"integer"},"tasksLimit":{"description":"Maximum number of tasks allowed","type":"integer"},"triggeredBy":{"description":"Units that triggered this service","type":"array","items":{"type":"string"}}}},"SystemdServiceList":{"type":"object","properties":{"count":{"type":"integer"},"services":{"type":"array","items":{"$ref":"#/definitions/SystemdService"}}}}}}